ade_parser/app/requirements.txt
Flask==3.0.0
ics==0.7.2
mysql-connector
pymongo==4.6.1
Requests==2.31.0
selenium==4.16.0
